Move utils to separate module
